Revert gunicorn/main.py
