+++ {{cookiecutter.project_slug}}/README.md
@@ -1,20 +1,9 @@
 {% set is_open_source = cookiecutter.open_source_license != 'Not open source' -%}
+
 # {{ cookiecutter.project_name }}
+
 Your main CLI program should be written `{{ cookiecutter.package_name }}.cli:cli`
 and will then be registered as the CLI command `{{ cookiecutter.project_slug }}` .
-
-{% if is_open_source %}
-.. image:: https://img.shields.io/pypi/v/{{ cookiecutter.project_slug }}.svg
-        :target: https://pypi.python.org/pypi/{{ cookiecutter.project_slug }}
-
-.. image:: https://readthedocs.org/projects/{{ cookiecutter.project_slug | replace("_", "-") }}/badge/?version=latest
-        :target: https://{{ cookiecutter.project_slug | replace("_", "-") }}.readthedocs.io/en/latest/?badge=latest
-        :alt: Documentation Status
-{%- endif %}
-
-.. image:: https://pyup.io/repos/github/{{ cookiecutter.github_username }}/{{ cookiecutter.project_slug }}/shield.svg
-     :target: https://pyup.io/repos/github/{{ cookiecutter.github_username }}/{{ cookiecutter.project_slug }}/
-     :alt: Updates
 
 
 {{ cookiecutter.project_short_description }}
