+++ {{cookiecutter.project_slug}}/conf/development.yaml
@@ -1,0 +1,5 @@
+---
+celery:
+  broker: 'redis://'
+  backend: 'redis://'
+  enable_utc: True
