+++ {{cookiecutter.project_slug}}/conf/gunicorn.py
@@ -1,0 +1,8 @@
+from {{cookiecutter.project_slug}}.config import logfile_path
+
+
+logconfig = logfile_path(debug=False)
+bind = "127.0.0.1:5000"
+workers = 4
+worker_class = "uvicorn.workers.UvicornWorker"
+preload_app = False
