+++ {{cookiecutter.project_slug}}/deploy/{{cookiecutter.project_slug}}.configmap.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+data:
+  config.yaml: |
+    {{ cookiecutter.project_slug }}:
+      token: changeme
+      debug: false
+      download_dir: /tmp/{{ cookiecutter.project_slug }}
+      env: development
+      url: https://{{ cookiecutter.project_slug }}.yourdomain.com
+    sentry:
+      url: null
+      environment: development
+kind: ConfigMap
+metadata:
+  name: {{cookiecutter.project_slug}}-config
