+++ {{cookiecutter.project_slug}}/setup.cfg
@@ -77,12 +77,17 @@ install_requires=
     aiohttp
     click
     fastapi[all]
-    mlflow
     pydantic
     sentry-sdk
     starlette
     starlette_exporter
     uvicorn
+    uvloop
+    fastapi
+    pydantic
+    PyYAML
+    requests
+    paramiko
 
 zip_safe=False
