+++ {{cookiecutter.project_slug}}/{{cookiecutter.package_name}}/api/{{cookiecutter.package_name}}.py
@@ -6,16 +6,16 @@ import logging
 from fastapi import APIRouter
 from pydantic import BaseModel, Field
 
+from {{cookiecutter.package_name}} import __version__
 
 router = APIRouter(prefix="/api/v1", tags=["{{cookiecutter.package_name}}"])
 
 logger = logging.getLogger(__name__)
 
 # CHANGE BELOW
-description = "A description of your App "\
-    "using string concatenation"
-title = "Custom App"
-version = "0.0.42"
+description = "{{project_short_description}}"
+title = "{{cookiecutter.project_name}"
+version = __version__
 
 
 class Item(BaseModel):
